added json functionality to client

+++ src/Server/header/base64.h
@@ -89,7 +89,7 @@ extern "C" {
 #endif
 
 int Base64encode_len(int len);
-int Base64encode(char * coded_dst, const char *plain_src,int len_plain_src);
+int Base64encode(char * coded_dst, const char *plain_src,size_t len_plain_src);
 
 int Base64decode_len(const char * coded_src);
 int Base64decode(char * plain_dst, const char *coded_src);
